evalengine: comparison benchmarks
