+++ go/vt/vtgate/evalengine/arithmetic_test.go
@@ -510,197 +510,6 @@ func TestNullSafeAdd(t *testing.T) {
 	}
 }
 
-func TestNullsafeCompare(t *testing.T) {
-	collation := collations.Local().LookupByName("utf8mb4_general_ci").ID()
-	tcases := []struct {
-		v1, v2 sqltypes.Value
-		out    int
-		err    error
-	}{{
-		// All nulls.
-		v1:  NULL,
-		v2:  NULL,
-		out: 0,
-	}, {
-		// LHS null.
-		v1:  NULL,
-		v2:  NewInt64(1),
-		out: -1,
-	}, {
-		// RHS null.
-		v1:  NewInt64(1),
-		v2:  NULL,
-		out: 1,
-	}, {
-		// LHS Text
-		v1:  TestValue(sqltypes.VarChar, "abcd"),
-		v2:  TestValue(sqltypes.VarChar, "abcd"),
-		out: 0,
-	}, {
-		// Make sure underlying error is returned for LHS.
-		v1:  TestValue(sqltypes.Int64, "1.2"),
-		v2:  NewInt64(2),
-		err: vterrors.New(vtrpcpb.Code_INVALID_ARGUMENT, "strconv.ParseInt: parsing \"1.2\": invalid syntax"),
-	}, {
-		// Make sure underlying error is returned for RHS.
-		v1:  NewInt64(2),
-		v2:  TestValue(sqltypes.Int64, "1.2"),
-		err: vterrors.New(vtrpcpb.Code_INVALID_ARGUMENT, "strconv.ParseInt: parsing \"1.2\": invalid syntax"),
-	}, {
-		// Numeric equal.
-		v1:  NewInt64(1),
-		v2:  NewUint64(1),
-		out: 0,
-	}, {
-		// Numeric unequal.
-		v1:  NewInt64(1),
-		v2:  NewUint64(2),
-		out: -1,
-	}, {
-		// Non-numeric equal
-		v1:  TestValue(sqltypes.VarBinary, "abcd"),
-		v2:  TestValue(sqltypes.Binary, "abcd"),
-		out: 0,
-	}, {
-		// Non-numeric unequal
-		v1:  TestValue(sqltypes.VarBinary, "abcd"),
-		v2:  TestValue(sqltypes.Binary, "bcde"),
-		out: -1,
-	}, {
-		// Date/Time types
-		v1:  TestValue(sqltypes.Datetime, "1000-01-01 00:00:00"),
-		v2:  TestValue(sqltypes.Binary, "1000-01-01 00:00:00"),
-		out: 0,
-	}, {
-		// Date/Time types
-		v1:  TestValue(sqltypes.Datetime, "2000-01-01 00:00:00"),
-		v2:  TestValue(sqltypes.Binary, "1000-01-01 00:00:00"),
-		out: 1,
-	}, {
-		// Date/Time types
-		v1:  TestValue(sqltypes.Datetime, "1000-01-01 00:00:00"),
-		v2:  TestValue(sqltypes.Binary, "2000-01-01 00:00:00"),
-		out: -1,
-	}, {
-		// Date/Time types
-		v1:  TestValue(sqltypes.Bit, "101"),
-		v2:  TestValue(sqltypes.Bit, "101"),
-		out: 0,
-	}, {
-		// Date/Time types
-		v1:  TestValue(sqltypes.Bit, "1"),
-		v2:  TestValue(sqltypes.Bit, "0"),
-		out: 1,
-	}, {
-		// Date/Time types
-		v1:  TestValue(sqltypes.Bit, "0"),
-		v2:  TestValue(sqltypes.Bit, "1"),
-		out: -1,
-	}}
-	for _, tcase := range tcases {
-		got, err := NullsafeCompare(tcase.v1, tcase.v2, collation)
-		if tcase.err != nil {
-			require.EqualError(t, err, tcase.err.Error())
-		}
-		if tcase.err != nil {
-			continue
-		}
-
-		if got != tcase.out {
-			t.Errorf("NullsafeCompare(%v, %v): %v, want %v", printValue(tcase.v1), printValue(tcase.v2), got, tcase.out)
-		}
-	}
-}
-
-func getCollationID(collation string) collations.ID {
-	id, _ := collations.Local().LookupID(collation)
-	return id
-}
-
-func TestNullsafeCompareCollate(t *testing.T) {
-	tcases := []struct {
-		v1, v2    string
-		collation collations.ID
-		out       int
-		err       error
-	}{
-		{
-			// case insensitive
-			v1:        "abCd",
-			v2:        "aBcd",
-			out:       0,
-			collation: getCollationID("utf8mb4_0900_as_ci"),
-		},
-		{
-			// accent sensitive
-			v1:        "ǍḄÇ",
-			v2:        "ÁḆĈ",
-			out:       1,
-			collation: getCollationID("utf8mb4_0900_as_ci"),
-		},
-		{
-			// hangul decomposition
-			v1:        "\uAC00",
-			v2:        "\u326E",
-			out:       0,
-			collation: getCollationID("utf8mb4_0900_as_ci"),
-		},
-		{
-			// kana sensitive
-			v1:        "\xE3\x81\xAB\xE3\x81\xBB\xE3\x82\x93\xE3\x81\x94",
-			v2:        "\xE3\x83\x8B\xE3\x83\x9B\xE3\x83\xB3\xE3\x82\xB4",
-			out:       -1,
-			collation: getCollationID("utf8mb4_ja_0900_as_cs_ks"),
-		},
-		{
-			// non breaking space
-			v1:        "abc ",
-			v2:        "abc\u00a0",
-			out:       -1,
-			collation: getCollationID("utf8mb4_0900_as_cs"),
-		},
-		{
-			// "cs" counts as a separate letter, where c < cs < d
-			v1:        "c",
-			v2:        "cs",
-			out:       -1,
-			collation: getCollationID("utf8mb4_hu_0900_ai_ci"),
-		},
-		{
-			v1:        "abcd",
-			v2:        "abcd",
-			collation: 0,
-			err:       vterrors.New(vtrpcpb.Code_UNKNOWN, "cannot compare strings, collation is unknown or unsupported (collation ID: 0)"),
-		},
-		{
-			v1:        "abcd",
-			v2:        "abcd",
-			collation: 1111,
-			err:       vterrors.New(vtrpcpb.Code_UNKNOWN, "cannot compare strings, collation is unknown or unsupported (collation ID: 1111)"),
-		},
-		{
-			v1: "abcd",
-			v2: "abcd",
-			// unsupported collation gb18030_bin
-			collation: 249,
-			err:       vterrors.New(vtrpcpb.Code_UNKNOWN, "cannot compare strings, collation is unknown or unsupported (collation ID: 249)"),
-		},
-	}
-	for _, tcase := range tcases {
-		got, err := NullsafeCompare(TestValue(sqltypes.VarChar, tcase.v1), TestValue(sqltypes.VarChar, tcase.v2), tcase.collation)
-		if !vterrors.Equals(err, tcase.err) {
-			t.Errorf("NullsafeCompare(%v, %v) error: %v, want %v", tcase.v1, tcase.v2, vterrors.Print(err), vterrors.Print(tcase.err))
-		}
-		if tcase.err != nil {
-			continue
-		}
-
-		if got != tcase.out {
-			t.Errorf("NullsafeCompare(%v, %v): %v, want %v", tcase.v1, tcase.v2, got, tcase.out)
-		}
-	}
-}
-
 func TestCast(t *testing.T) {
 	tcases := []struct {
 		typ sqltypes.Type
